# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CodeMind AI API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/organizations/{org_id}/pr-reviews": {
      "post": {
        "tags": [
          "pr-review"
        ],
        "summary": "Review Pull Request",
        "operationId": "review_pull_request_api_organizations__org_id__pr_reviews_post",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Org Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartPRReviewRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PRReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "pr-review"
        ],
        "summary": "List Pr Reviews",
        "operationId": "list_pr_reviews_api_organizations__org_id__pr_reviews_get",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Org Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PRReviewResponse"
                  },
                  "title": "Response List Pr Reviews Api Organizations  Org Id  Pr Reviews Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/organizations/{org_id}/pr-reviews/{pr_review_id}": {
      "get": {
        "tags": [
          "pr-review"
        ],
        "summary": "Get Pr Review",
        "operationId": "get_pr_review_api_organizations__org_id__pr_reviews__pr_review_id__get",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Org Id"
            }
          },
          {
            "name": "pr_review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pr Review Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PRReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PRReviewResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "owner": {
            "type": "string",
            "title": "Owner"
          },
          "repo": {
            "type": "string",
            "title": "Repo"
          },
          "pr_number": {
            "type": "integer",
            "title": "Pr Number"
          },
          "pr_url": {
            "type": "string",
            "title": "Pr Url"
          },
          "commit_sha": {
            "type": "string",
            "title": "Commit Sha"
          },
          "findings_count": {
            "type": "integer",
            "title": "Findings Count"
          },
          "comments_posted": {
            "type": "integer",
            "title": "Comments Posted"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "review_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "owner",
          "repo",
          "pr_number",
          "pr_url",
          "commit_sha",
          "findings_count",
          "comments_posted",
          "status",
          "review_url",
          "created_at"
        ],
        "title": "PRReviewResponse"
      },
...
      "StartPRReviewRequest": {
        "properties": {
          "pr_number": {
            "type": "integer",
            "title": "Pr Number"
          }
        },
        "type": "object",
        "required": [
          "pr_number"
        ],
        "title": "StartPRReviewRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}